refactor(router): share tolerant scalar facts
