+++ sglang_omni_router/rust/src/speech_facts.rs
@@ -219,18 +219,31 @@ pub(crate) fn named_voice(fields: &SpeechFields, references: &[ReferenceForm]) -
             .is_some_and(|voice| !voice.is_empty() && !voice.eq_ignore_ascii_case("default"))
 }
 
-enum ScalarFact<'a> {
+pub(crate) enum ScalarFact<'a> {
     String(Cow<'a, str>),
+    Bool(bool),
+    Signed(i64),
+    Unsigned(u64),
+    Float(f64),
     Null,
     Other,
 }
 
 impl ScalarFact<'_> {
+    pub(crate) fn into_string(self) -> Option<String> {
+        match self {
+            Self::String(value) => Some(value.into_owned()),
+            _ => None,
+        }
+    }
+
     fn into_nullable_string(self) -> (Option<String>, bool) {
         match self {
             Self::String(value) => (Some(value.into_owned()), true),
             Self::Null => (None, true),
-            Self::Other => (None, false),
+            Self::Bool(_) | Self::Signed(_) | Self::Unsigned(_) | Self::Float(_) | Self::Other => {
+                (None, false)
+            }
         }
     }
 
@@ -238,12 +251,26 @@ impl ScalarFact<'_> {
         match self {
             Self::String(_) => (true, true),
             Self::Null => (false, true),
-            Self::Other => (false, false),
+            Self::Bool(_) | Self::Signed(_) | Self::Unsigned(_) | Self::Float(_) | Self::Other => {
+                (false, false)
+            }
+        }
+    }
+
+    pub(crate) fn into_bool(self) -> Option<bool> {
+        match self {
+            Self::Bool(value) => Some(value),
+            Self::Signed(value) => bool_from_integer(value),
+            Self::Unsigned(value) => bool_from_integer(value),
+            Self::Float(0.0) => Some(false),
+            Self::Float(1.0) => Some(true),
+            Self::String(value) => parse_bool_fact(&value),
+            Self::Float(_) | Self::Null | Self::Other => None,
         }
     }
 }
 
-struct ScalarFactSeed;
+pub(crate) struct ScalarFactSeed;
 
 impl<'de> DeserializeSeed<'de> for ScalarFactSeed {
     type Value = ScalarFact<'de>;
@@ -280,20 +307,20 @@ impl<'de> DeserializeSeed<'de> for ScalarFactSeed {
                 Ok(ScalarFact::Null)
             }
 
-            fn visit_bool<E>(self, _value: bool) -> Result<Self::Value, E> {
-                Ok(ScalarFact::Other)
+            fn visit_bool<E>(self, value: bool) -> Result<Self::Value, E> {
+                Ok(ScalarFact::Bool(value))
             }
 
-            fn visit_i64<E>(self, _value: i64) -> Result<Self::Value, E> {
-                Ok(ScalarFact::Other)
+            fn visit_i64<E>(self, value: i64) -> Result<Self::Value, E> {
+                Ok(ScalarFact::Signed(value))
             }
 
-            fn visit_u64<E>(self, _value: u64) -> Result<Self::Value, E> {
-                Ok(ScalarFact::Other)
+            fn visit_u64<E>(self, value: u64) -> Result<Self::Value, E> {
+                Ok(ScalarFact::Unsigned(value))
             }
 
-            fn visit_f64<E>(self, _value: f64) -> Result<Self::Value, E> {
-                Ok(ScalarFact::Other)
+            fn visit_f64<E>(self, value: f64) -> Result<Self::Value, E> {
+                Ok(ScalarFact::Float(value))
             }
 
             fn visit_seq<A>(self, sequence: A) -> Result<Self::Value, A::Error>
@@ -313,6 +340,35 @@ impl<'de> DeserializeSeed<'de> for ScalarFactSeed {
             }
         }
         deserializer.deserialize_any(ScalarVisitor)
+    }
+}
+
+fn bool_from_integer<T>(value: T) -> Option<bool>
+where
+    T: Eq + From<u8>,
+{
+    if value == T::from(0) {
+        Some(false)
+    } else if value == T::from(1) {
+        Some(true)
+    } else {
+        None
+    }
+}
+
+fn parse_bool_fact(value: &str) -> Option<bool> {
+    if ["1", "on", "t", "true", "y", "yes"]
+        .iter()
+        .any(|candidate| value.eq_ignore_ascii_case(candidate))
+    {
+        Some(true)
+    } else if ["0", "off", "f", "false", "n", "no"]
+        .iter()
+        .any(|candidate| value.eq_ignore_ascii_case(candidate))
+    {
+        Some(false)
+    } else {
+        None
     }
 }
